代码积累

+++ code-interview/第9章 其他题目/9-设计LRU缓存结构.cpp
@@ -49,7 +49,7 @@ void jset(int a, int b){
   head=t, mp[a]=b, mn[a]=t, sz++;
 };
 
-int jget(int a){
+long long jget(int a){
   map<long long, long long>::iterator p=mp.find(a);
   if(p==mp.end()) return -1;
   jnew(mn.find(a)->second);
